Extract company service tier
