+++ apps/web/app/api/v1/taxonomies/route.ts
@@ -3,13 +3,13 @@ import { type NextRequest, NextResponse } from "next/server";
 // rather than the `"use server"` action modules (`@/lib/actions/*`). The
 // service functions are functionally identical but avoid the
 // server-action machinery (per-call RPC URL, serialization boundary,
-// security IDs). See issues #3231 / #3329.
+// security IDs). See issues #3231 / #3329 / #3331.
 import {
   getAllSeniorities,
   getAllOccupationsGrouped,
   getAllTechnologiesGrouped,
 } from "@/lib/services/taxonomy";
-import { suggestIndustries } from "@/lib/actions/company";
+import { suggestIndustries } from "@/lib/services/company";
 import { CACHE_TTL_LONG } from "@/lib/cache-ttl";
 import { checkRateLimit, apiResponse } from "../_shared";
